✨ Feat: selectedTabItem분리
